+++ tests/scalers/redis/redis_sentinel_strings/redis_sentinel_strings_test.go
@@ -1,0 +1,211 @@
+//go:build e2e
+// +build e2e
+
+package redis_standalone_keyvalue_test
+
+import (
+	"encoding/base64"
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/joho/godotenv"
+	"github.com/stretchr/testify/assert"
+	"k8s.io/client-go/kubernetes"
+
+	. "github.com/kedacore/keda/v2/tests/helper"
+	redis "github.com/kedacore/keda/v2/tests/scalers/redis/helper"
+)
+
+// Load environment variables from .env file
+var _ = godotenv.Load("../../.env")
+
+const (
+	testName = "redis-sentinel-keyvalue-test"
+)
+
+var (
+	testNamespace             = fmt.Sprintf("%s-ns", testName)
+	redisNamespace            = fmt.Sprintf("%s-redis-ns", testName)
+	deploymentName            = fmt.Sprintf("%s-deployment", testName)
+	clientName                = fmt.Sprintf("%s-client", testName)
+	scaledObjectName          = fmt.Sprintf("%s-so", testName)
+	triggerAuthenticationName = fmt.Sprintf("%s-ta", testName)
+	secretName                = fmt.Sprintf("%s-secret", testName)
+	redisPassword             = "admin"
+	redisKey                  = fmt.Sprintf("%s-key", testName)
+	redisHost                 = fmt.Sprintf("%s-headless", testName)
+	minReplicaCount           = 0
+	maxReplicaCount           = 2
+)
+
+type templateData struct {
+	TestNamespace             string
+	RedisNamespace            string
+	DeploymentName            string
+	ScaledObjectName          string
+	TriggerAuthenticationName string
+	SecretName                string
+	MinReplicaCount           int
+	MaxReplicaCount           int
+	RedisPassword             string
+	RedisPasswordBase64       string
+	RedisKey                  string
+	RedisHost                 string
+}
+
+const (
+	deploymentTemplate = `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: {{.DeploymentName}}
+  namespace: {{.TestNamespace}}
+spec:
+  selector:
+    matchLabels:
+      app: {{.DeploymentName}}
+  replicas: 0
+  template:
+    metadata:
+      labels:
+        app: {{.DeploymentName}}
+    spec:
+      containers:
+      - name: my-app
+        image: nginxinc/nginx-unprivileged
+        imagePullPolicy: IfNotPresent
+        ports:
+        - containerPort: 80
+        env:
+        - name: REDIS_ADDRESSES
+          value: {{.RedisHost}}.{{.RedisNamespace}}:26379
+        - name: REDIS_PORT
+          value: "6379"
+`
+
+	secretTemplate = `apiVersion: v1
+kind: Secret
+metadata:
+  name: {{.SecretName}}
+  namespace: {{.TestNamespace}}
+type: Opaque
+data:
+  password: {{.RedisPasswordBase64}}
+`
+
+	triggerAuthenticationTemplate = `apiVersion: keda.sh/v1alpha1
+kind: TriggerAuthentication
+metadata:
+  name: {{.TriggerAuthenticationName}}
+  namespace: {{.TestNamespace}}
+spec:
+  secretTargetRef:
+  - parameter: password
+    name: {{.SecretName}}
+    key: password
+  - parameter: sentinelPassword
+    name: {{.SecretName}}
+    key: password
+`
+
+	scaledObjectTemplate = `apiVersion: keda.sh/v1alpha1
+kind: ScaledObject
+metadata:
+  name: {{.ScaledObjectName}}
+  namespace: {{.TestNamespace}}
+spec:
+  scaleTargetRef:
+    name: {{.DeploymentName}}
+  pollingInterval: 5
+  cooldownPeriod:  10
+  minReplicaCount: {{.MinReplicaCount}}
+  maxReplicaCount: {{.MaxReplicaCount}}
+  triggers:
+  - type: redis
+    metadata:
+      addressesFromEnv: REDIS_ADDRESSES
+      sentinelMaster: mymaster
+      keyName: {{.RedisKey}}
+      keyValue: "1.5"
+      activationKeyValue: "5"
+    authenticationRef:
+      name: {{.TriggerAuthenticationName}}
+`
+)
+
+func TestScaler(t *testing.T) {
+	// Create kubernetes resources for PostgreSQL server
+	kc := GetKubernetesClient(t)
+	data, templates := getTemplateData()
+	t.Cleanup(func() {
+		redis.RemoveSentinel(t, testName, redisNamespace)
+		redis.RemoveClient(t, clientName, redisNamespace)
+		DeleteKubernetesResources(t, testNamespace, data, templates)
+	})
+
+	// Create Redis Standalone
+	redis.InstallSentinel(t, kc, testName, redisNamespace, redisPassword)
+	redis.InstallClient(t, kc, clientName, redisNamespace)
+	// wait until client is ready
+	time.Sleep(10 * time.Second)
+
+	// Create kubernetes resources for testing
+	CreateKubernetesResources(t, kc, testNamespace, data, templates)
+
+	testActivation(t, kc, data)
+	testScaleOut(t, kc, data)
+	testScaleIn(t, kc)
+}
+
+func testActivation(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+	t.Log("--- testing activation ---")
+	setKeyValue(t, 4)
+
+	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, minReplicaCount, 60)
+}
+
+func testScaleOut(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+	t.Log("--- testing scale out ---")
+	setKeyValue(t, 10)
+
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, maxReplicaCount, 60, 3),
+		"replica count should be %d after 3 minutes", maxReplicaCount)
+}
+
+func testScaleIn(t *testing.T, kc *kubernetes.Clientset) {
+	t.Log("--- testing scale in ---")
+	setKeyValue(t, 0)
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, minReplicaCount, 60, 3),
+		"replica count should be %d after 3 minutes", minReplicaCount)
+}
+
+var data = templateData{
+	TestNamespace:             testNamespace,
+	RedisNamespace:            redisNamespace,
+	DeploymentName:            deploymentName,
+	ScaledObjectName:          scaledObjectName,
+	MinReplicaCount:           minReplicaCount,
+	MaxReplicaCount:           maxReplicaCount,
+	TriggerAuthenticationName: triggerAuthenticationName,
+	SecretName:                secretName,
+	RedisPassword:             redisPassword,
+	RedisPasswordBase64:       base64.StdEncoding.EncodeToString([]byte(redisPassword)),
+	RedisKey:                  redisKey,
+	RedisHost:                 redisHost,
+}
+
+func getTemplateData() (templateData, []Template) {
+	return data, []Template{
+		{Name: "secretTemplate", Config: secretTemplate},
+		{Name: "deploymentTemplate", Config: deploymentTemplate},
+		{Name: "triggerAuthenticationTemplate", Config: triggerAuthenticationTemplate},
+		{Name: "scaledObjectTemplate", Config: scaledObjectTemplate},
+	}
+}
+
+func setKeyValue(t *testing.T, value int) {
+	_, _, err := ExecCommandOnSpecificPod(t, clientName, redisNamespace,
+		fmt.Sprintf(`redis-cli --pass %s -h %s -p 26379 SENTINEL get-master-addr-by-name mymaster | xargs -n 2 sh -c 'redis-cli --pass %s -h $0 -p $1 SET %s %d'`,
+			redisPassword, redisHost, redisPassword, redisKey, value))
+	assert.NoErrorf(t, err, "cannot execute command - %s", err)
+}